Push final full-web view.
Without paying for custom domain hosting with the possibility of voiding some sort of Trademark or Copyright, This project must be converted into a Browser Extension. I will be working on getting this fully functional as a browser extension as I continue working on this project.
Remember, I am not profiting in any way from this and my only intention for this is to use it as a learning experience. If you want to help and support me, feel free to donate using the funding links provided.